// packages/blocks/ordermanagement/src/types.ts
export interface TrackingDetail {
  id: string;
  type: string;
  attributes: TrackingDetailAttributes;
}
interface TrackingDetailAttributes {
  shipment_provider_name: string;
  tracking_number: string;
  additional_notes?: string | null;
  added_by: string;
  added_by_email: string;
  added_date: string;
}